feat(cli): Implement join(just login) and register commands
